% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.garthwaitecov}
\alias{.garthwaitecov}
\title{Garthwaite-Covariance matrix}
\usage{
.garthwaitecov(mat)
}
\arguments{
\item{A, }{Any matrix that you wish to take the square root of}
}
\description{
Calculate gartwaite-covariance matrix
}
\examples{
A1 <- matrix(2,1,1)
.garthwaitecov(A1)
A3 <- matrix(c(2,0.1,0.3,0.1,1,0.5,0.1,0.2,3),3,3)
.garthwaitecov(A3)

}
