% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.mpinv}
\alias{.mpinv}
\title{Helper function for pseudoinverse}
\usage{
.mpinv(A, eps = 1e-13)
}
\arguments{
\item{A, }{Any matrix that you wish to invert}

\item{eps}{optional, the level of tolerance to use}
}
\description{
Calculate the Penrose-Moore pseudoinverse based on adjusting the eigenvalues of the matrix
}
\examples{
A1 <- matrix(2,1,1)
.mpinv(A1)
A2 <- 3
.mpinv(A2)
A3 <- matrix(c(2,0.1,0.3,0.1,1,0.5),2,3)
.mpinv(A3)
round(A3\%*\%.mpinv(A3),2)

}
