% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\docType{data}
\name{factors}
\alias{factors}
\title{Fama-French factors: value and operating profitability belonging to X and Y}
\format{An xts monthly timeseries with yearmon time and 25 columns}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
factors
}
\description{
Data was accessed on 2018-09-15 and saved to xts file
}
\keyword{datasets}
