% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISminimax_one.R
\name{ISminimax_one}
\alias{ISminimax_one}
\title{Calculate (in-sample) the coefficient to adjust the optimal Markowitz portfolio rule based on financial turbulence}
\usage{
ISminimax_one(X, weights = NULL, method = NULL, s.k = 1,
  imp = FALSE, rf = TRUE)
}
\arguments{
\item{X}{the dataset X in xts format (if not, the program will try to convert it to xts). Monthly returns.}

\item{weights}{optional, can either be a vector (in which case it will be used for each row and must match ncol(X)) or a matrix with the same dimension as X
if weights=NULL (the default), equal weights will be generated based on the number of columns (CHECK no of non NA entries per row)}

\item{method}{optional, if a robust mean/covariance estimator should be used}

\item{s.k}{optional, lookback window for current observations that are related to the long-term mean}

\item{imp}{optional, should missing values be imputed? (standard: FALSE)}

\item{rf}{optional, should the case with (default) or without a risk-free rate be used (the latter case is not implemented)}
}
\value{
timeseries/vector
}
\description{
Adjust the optimal Markowitz Portfolio rule according to model 1.2.2 of Garlappi, Uppal & Wang (2008, RFS)
According tom the model, we use the minimax rule of Gilboa & Schmeidler (1989) to use the worst (min/inf) of all possible
priors for the calculation of the optimal portfolio rule.
Uncertainty about expected returns estimated jointly for all assets
This coefficient explains how much to move out of the provided factor towards the risk-free rate or (if using the model without the risk-free rate)
towards the global Minimum-Variance portfolio
The measure of current uncertainty is based on Stöckl (2019, WP)
}
\examples{
require(xts)
data(X)
sturb1 <- ISminimax_one(X, s.k=1)
sturb2 <- ISminimax_one(X,s.k=12)
sturb3 <- ISminimax_one(X,s.k=12, method="MCD")
plot(sturb1$sturb, lwd=2, main="Portfolio Adjustment coefficient with risk-free rate",ylim=c(-0.1,1))
lines(sturb2$sturb,lwd=2,col="red", main="Portfolio Adjustment coefficient based on turbulence with 12 month lookback period",on=NA,ylim=c(-0.1,1))
lines(sturb3$sturb,lwd=2,col="blue", main="Portfolio Adjustment coefficient based on turbulence with 12 month lookback period and robust MCD estimator",on=NA,ylim=c(-0.1,1))

}
