% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISturbulence.R
\name{ISturbulence}
\alias{ISturbulence}
\title{Calculate In-sample turbulence}
\usage{
ISturbulence(X, weights = NULL, squared = FALSE, norm = FALSE,
  method = NULL, s.k = 1, imp = FALSE, use = "complete.obs")
}
\arguments{
\item{X}{the dataset X in xts format (if not, the program will try to convert it to xts)}

\item{weights}{optional, can either be a vector (in which case it will be used for each row and must match ncol(X)) or a matrix with the same dimension as X
if weights=NULL (the default), equal weights will be generated based on the number of columns (CHECK no of non NA entries per row)}

\item{squared}{optional, should the square root be taken of the results? (standard: TRUE which correwsponds to squared=FALSE)}

\item{norm}{optional, should the result be normalized by 1/sum(weights^2) per row (makes expectation equal to 1)}

\item{method}{optional, if a robust mean/covariance estimator should be used}

\item{s.k}{optional, lookback window for current observations that are related to the long-term mean}

\item{imp}{optional, should missing values be imputed? (standard:)}

\item{use}{optional, use="pairwise.complete.obs": What method of dealing with missing values should be used (note, that variables
with more than half missing obs are thrown out anyway)}
}
\value{
list containing 4 elements:
   - turb: turbulence index
}
\description{
Calculate IS-turbulence based on a cross-section of times-series (xts) data
}
\examples{
require(xts)
data(X)
turb1 <- ISturbulence(X, s.k=1)$turb
turb2 <- ISturbulence(X,s.k=12)$turb
turb3 <- ISturbulence(X,s.k=12, method="MCD")$turb
plot(turb1, lwd=2, main="In-sample turbulence (standard)")
lines(turb2,lwd=2,col="red", main="Turbulence with 12 month lookback period",on=NA)
lines(turb3,lwd=2,col="blue", main="Turbulence with 12 month lookback period and robust MCD estimator",on=NA)

}
