% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.R
\docType{data}
\name{X}
\alias{X}
\title{Fama-French 25 portfolios sorted according to size and value}
\format{An xts monthly timeseries with yearmon time and 25 columns}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
X
}
\description{
Data was accessed on 2018-09-15 and saved to xts file
}
\keyword{datasets}
