% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Y.R
\docType{data}
\name{Y}
\alias{Y}
\title{Fama-French 25 portfolios sorted according to size and operating profitability}
\format{An xts monthly timeseries with yearmon time and 25 columns}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
Y
}
\description{
Data was accessed on 2018-09-15 and saved to xts file
}
\keyword{datasets}
