% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.garthwaitecov}
\alias{.garthwaitecov}
\title{Helper function for maxtrix root}
\usage{
.garthwaitecov(mat)
}
\arguments{
\item{A, }{Any matrix that you wish to take the square root of}
}
\description{
Calculate matrix root according to https://www.arpm.co/lab/matrix-transpose-square-root.html#x350-152200050.8.2
}
\examples{
A1 <- matrix(2,1,1)
.ricattiRoot(A1)
A2 <- 3
.ricattiRoot(A2)
A3 <- matrix(c(2,0.1,0.3,0.1,1,0.5,0.1,0.2,3),3,3)
.garthwaitecov(A3)

}
